[FEAT] 포인트 히스토리 보기 구현
