--- src/main/java/kr/hhplus/be/server/interfaces/api/point_history/PointHistoryResponse.java
@@ -1,3 +1,0 @@
-package kr.hhplus.be.server.interfaces.api.point_history;
-
-public record PointHistoryResponse(Long userId, int amount, String type) {}
