+++ src/test/java/kr/hhplus/be/server/interfaces/api/point/PointController.java
@@ -2,9 +2,13 @@ package kr.hhplus.be.server.interfaces.api.point;
 
 import kr.hhplus.be.server.application.common.ApiResponse;
 import kr.hhplus.be.server.application.point.PointService;
+import kr.hhplus.be.server.domain.point.PointHistory;
 import kr.hhplus.be.server.domain.point.UserPoint;
+import kr.hhplus.be.server.interfaces.api.pointhistory.PointHistoryResponse;
 import lombok.RequiredArgsConstructor;
 import org.springframework.web.bind.annotation.*;
+
+import java.util.List;
 
 @RestController
 @RequestMapping("/api/points")
@@ -35,5 +39,14 @@ public class PointController {
         PointResponse pointResponse = PointResponse.from(userPoint);
 
         return ApiResponse.success(pointResponse);
+    }
+
+    @GetMapping("/{userId}")
+    public List<PointHistoryResponse> getHistory(@PathVariable Long userId) {
+        List<PointHistory> pointHistoryList = pointService.getPointHistory(userId);
+
+        return pointHistoryList.stream()
+                .map(PointHistoryResponse::fromEntity)
+                .toList();
     }
 }
